+++ src/main.js
@@ -55,3 +55,4 @@ export { default as TextField } from './components/TextField';
 export { default as Toolbar } from './components/Toolbar';
 export { default as Token } from './components/Token';
 export { default as TokensInput } from './components/TokensInput';
+export { default as NavBar } from './components/NavBar';